==> 2_concate_codes.ipynb <==
import pandas as pd

kg_frames= []
for i in range(1,51):
    f_name = "D:/Th/Data/KGT_with_no_output/Codes/code_batch_" + str(i)  + "_no_output.csv"
    kg_data = pd.read_csv(f_name, dtype='unicode')
    kg_frames.append(kg_data)
    del kg_data

kg_codes = pd.concat(kg_frames)
del kg_frames

kg_codes.drop(['Unnamed: 0'], axis = 1, inplace=True)

kg_codes.columns

kg_codes.to_csv('D:/Th/Data/KGT_with_no_output/KGT_codes.csv')

del kg_codes

kg_frames= []
for i in range(1,51):
    try:
        f_name = "D:/Th/Data/KGT_with_no_output/Markdowns/markdown_batch_" + str(i) + ".csv"
        kg_md = pd.read_csv(f_name, dtype='unicode')
        kg_frames.append(kg_md)
    except:
        pass #part 36 not exist!

kg_md = pd.concat(kg_frames)
del kg_frames

kg_md.columns

kg_md.to_csv('D:/Th/data/KGT_with_no_output/KGT_Markdown.csv')
