==> 4_notebook_metrics.ipynb <==
import pandas as pd
from tqdm import tqdm
tqdm.pandas()

### aggregate code cells

code_df = pd.read_csv('D:/Th/data/KGT_metrics/Code_cell_metrics.csv')

code_df.head()

#cell_metrics.project_ID = cell_metrics.project_ID.astype(int)

pm1 = code_df[['project_ID','LOC', 'BLC', 'UDF', 'I', 'EH','NDD','NEC','S','P'
               ,'OPRND','OPRATOR','UOPRND','UOPRATOR','ID','LOCom', 'EAP',
               'CW']].groupby(by='project_ID').sum().reset_index()

pm2 =  code_df[['project_ID','ALID','LOC','ALLC','KLCID','CyC']].groupby(by='project_ID').mean().reset_index()
pm2.rename(columns={'LOC': 'MeanLCC'}, inplace= True)

pm3 =  code_df[['project_ID','MLID','NBD','ID']].groupby(by='project_ID').max().reset_index().rename(columns={'ID':'AID'})

pm4 = code_df[['project_ID', 'LOC']].groupby('project_ID').count().reset_index().rename(columns={'LOC':'CC'})

notebook_cell1 = pd.concat([pm1, pm2.drop(columns=['project_ID']), 
                            pm3.drop(columns=['project_ID']), pm4.drop(columns=['project_ID'])], axis=1)

### aggregate markdown cells

md_df = pd.read_csv('D:/Th/data/KGT_metrics/markdown_cell_metrics.csv')

qm1 = md_df[['project_ID', 'H1']].groupby('project_ID').count().reset_index().rename(columns={'H1': 'MC'})

qm2 = md_df[['project_ID', 'MW', 'LMC']].groupby('project_ID').mean().reset_index().rename(columns={'MW':'MeanWMC', 'LMC':'MeanLMC'})

md_df.columns

qm3 = md_df[['project_ID', 'H1', 'H2', 'H3', 'MW', 'LMC']].groupby('project_ID').sum().reset_index()

notebook_cell2 = pd.concat([qm1, qm2.drop(columns=['project_ID']), qm3.drop(columns=['project_ID'])], axis=1)

notebook_cell1.head()

clean notebook cell metrics data

notebook_cell1 = notebook_cell1[notebook_cell1.project_ID != 'train_emb']
notebook_cell2 = notebook_cell2[notebook_cell1.project_ID != 'train_emb']
notebook_cell1.project_ID =notebook_cell1['project_ID'].astype(int)
notebook_cell2.project_ID =notebook_cell2['project_ID'].astype(int)
notebook_cell1 = notebook_cell1[notebook_cell1['LOC'] > 1]

## load performance tier of notebook developers

user_pt = pd.read_csv('D:/Th/data/user_pt.csv').rename(columns={'Id_x': 'project_ID', 
                                                               'PerformanceTier':'PT'}).drop(columns=['Unnamed: 0'])

notebook_cell = notebook_cell1.merge(notebook_cell2, how='left').merge(user_pt, how='left').fillna(0)

notebook_cell.columns

notebook_cell.to_csv('D:/Th/data/KGT_metrics/notebook_metrics.csv')

notebook_cell.shape
